qr_generator/requirements.txt
Pillow==6.2.0
qrcode==6.1
six==1.12.0
cairosvg